refactor: 클래스 파일을 분리
