+++ app/src/main/java/woowacourse/shopping/data/cart/model/CartPageData.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.data.cart.model
+
+data class CartPageData(
+    val cartItems: List<CartItemData>,
+    val pageNumber: Int,
+    val totalPageSize: Int,
+    val pageSize: Int,
+    val totalProductSize: Int,
+)
